[ELY-2925] Update JwtTestUtil so 'use' or 'key_ops' parameters can be specified.
Also split out JWK specific functionality to JwkTestUtil so it can be
used without needing the okhttp3.mockwebserver.Dispatcher on the
classpath at runtime.
